slight change to the file size formatting function

+++ macosx/NSStringAdditions.m
@@ -49,7 +49,7 @@
 + (NSString *) stringForFileSize: (uint64_t) size
 {
     const CGFloat baseFloat = [NSApp isOnSnowLeopardOrBetter] ? 1000.0 : 1024.0;
-    const NSInteger baseInt = [NSApp isOnSnowLeopardOrBetter] ? 1000 : 1024;
+    const NSUInteger baseInt = [NSApp isOnSnowLeopardOrBetter] ? 1000 : 1024;
     
     if (size < baseInt)
     {
@@ -68,17 +68,17 @@
     }
     else if (size < pow(baseInt, 3))
     {
-        convertedSize = size / pow(baseFloat, 2);
+        convertedSize = size / powf(baseFloat, 2);
         unit = NSLocalizedString(@"MB", "File size - megabytes");
     }
     else if (size < pow(baseInt, 4))
     {
-        convertedSize = size / pow(baseFloat, 3);
+        convertedSize = size / powf(baseFloat, 3);
         unit = NSLocalizedString(@"GB", "File size - gigabytes");
     }
     else
     {
-        convertedSize = size / pow(baseFloat, 4);
+        convertedSize = size / powf(baseFloat, 4);
         unit = NSLocalizedString(@"TB", "File size - terabytes");
     }
